+++ src/lib/FormValue.ts
@@ -1,20 +1,20 @@
 class FormValue {
   static $name = 'FormValue'
   static clearValidate = function(_formValue: FormValue, ..._args: any[]) { console.error('未定义clearValidate函数') }
-  static validate = function(_formValue: FormValue, ..._args: any[]): Promise<unknown> {
+  static validate = function(_formValue: FormValue, ..._args: any[]): Promise<any> {
     console.error('未定义validate函数')
     return Promise.reject({ status: 'fail', code: 'undefined validate function' })
   }
-  ref: unknown
-  data: Record<PropertyKey, unknown>
+  ref: any
+  data: Record<PropertyKey, any>
   constructor() {
     this.ref = null
     this.data = {}
   }
-  setRef(ref: unknown) {
+  setRef(ref: any) {
     this.ref = ref
   }
-  setData(data: Record<PropertyKey, unknown>) {
+  setData(data: Record<PropertyKey, any>) {
     this.data = data
   }
   getRef() {
@@ -23,10 +23,10 @@ class FormValue {
   getData() {
     return this.data
   }
-  clearValidate(...args: unknown[]): void {
+  clearValidate(...args: any[]): void {
     return (this.constructor as typeof FormValue).clearValidate(this, ...args)
   }
-  validate(...args: unknown[]): Promise<unknown> {
+  validate(...args: any[]): Promise<any> {
     return (this.constructor as typeof FormValue).validate(this, ...args)
   }
 }